pyproject.toml
[project]
name = "cu-removed-hindi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-contentunderstanding>=1.0.1",
    "azure-core>=1.39.0",
    "azure-identity>=1.25.3",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
